Add 'done' flag to track successful icon creation
Introduces a 'done' boolean field in DiskIcons and its subclasses to indicate whether icon creation and writing to disk succeeded. Updates logic in IcnsHeader, IcoHeader, and PngHeader to set this flag upon successful completion. Modifies Icons and ListItem to use this flag for more accurate status reporting when creating or saving icon files.